requirements.txt
streamlit>=1.32.0
openai>=1.12.0
anthropic>=0.18.1
google-generativeai>=0.3.2
playwright>=1.41.0
beautifulsoup4>=4.12.0
pandas>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.0
duckduckgo-search>=4.5.0
readability-lxml>=0.8.1
pytest==7.4.3
pytest-cov==4.1.0
lxml[html_clean]==4.9.3
pydantic
openpyxl 